requirements.txt
deepgram-sdk
python-dotenv
tiktoken
ollama
psycopg2-binary
python-dotenv